+++ packages/expo-notifications/src/TopicSubscribeModule.android.ts
@@ -1,0 +1,5 @@
+import { NativeModulesProxy } from 'expo-modules-core';
+
+import { TopicSubscribeModule } from './TopicSubscribeModule.types';
+
+export default NativeModulesProxy.ExpoTopicSubscribeModule as any as TopicSubscribeModule;