+++ packages/expo-notifications/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unpublished
 
+- [android] Support subscribing to Firebase broadcast topics by ([#17674](https://github.com/expo/expo/pull/17674) [@mmomtchev](https://github.com/mmomtchev))
+
 ### 🛠 Breaking changes
 
 - [android] Set the "notification number" (sometimes used to increment badge count on some launchers) from the notification payload `badge` field. ([#17171](https://github.com/expo/expo/pull/17171) by [@danstepanov](https://github.com/danstepanov))